+++ tests/integration/cases/int_list_single/FSharp_sequence_array_decl_let_mutable@v8.fs
@@ -1,0 +1,8 @@
+module Main
+
+type Val =
+    | FInt of int64
+    | FList of Val list
+let mutable my_data: Val array = [|
+    FInt 1L
+|]