+++ tests/integration/cases/int_list_single/Java_combined@jdk_11.java
@@ -1,0 +1,10 @@
+class Main {
+    public static void main() {
+var my_data = new int[]{
+    1
+};
+my_data = new int[]{
+    1
+};
+    }
+}