+++ tests/integration/cases/int_list_single/Lua@v5_4.lua
@@ -1,0 +1,3 @@
+local my_data = {
+    1,
+}