+++ tests/integration/cases/int_list_single/Scala_combined@v3.scala
@@ -1,0 +1,8 @@
+object Fixture_int_list_single_Scala_combined {
+var my_data = List[Int](
+    1,
+)
+my_data = List[Int](
+    1,
+)
+}