+++ tests/integration/cases/int_list_single/Scala_sequence_seq_decl_var@v3.scala
@@ -1,0 +1,5 @@
+object Fixture_int_list_single_Scala_sequence_seq_decl_var {
+var my_data = Seq(
+    1,
+)
+}