+++ tests/integration/cases/int_list_single/V_combined@v0_4.v
@@ -1,0 +1,10 @@
+
+fn main() {
+	mut my_data := [
+		1,
+	]
+	my_data = [
+		1,
+	]
+	_ = my_data
+}