+++ tests/integration/cases/int_list_single/Zig_combined@v0_12.zig
@@ -1,0 +1,21 @@
+const ZVal = union(enum) {
+    nil,
+    bool: bool,
+    int: i64,
+    uint: u64,
+    float: f64,
+    str: []const u8,
+    arr: []const ZVal,
+    map: []const ZKV,
+    set: []const ZVal,
+};
+const ZKV = struct { key: []const u8, val: ZVal };
+pub fn main() void {
+    var my_data: ZVal = .{ .arr = &.{
+        .{ .int = 1 },
+    }};
+    my_data = .{ .arr = &.{
+        .{ .int = 1 },
+    }};
+    my_data = .nil;
+}